refactor: move to FriendshipHistoryRepository
